Implemented a custom cursor.... However..
When regulating the FPS, this clearly shows an issue with our current method of handling framerate. You can see the custom cursor lag behind SDL's seperate cursor.

The issue is that we regulate the entire game loop, when we should only be regulating the physics engine.. Hopefully the two are coded correctly so we can split them up and regulate the framerate properly